requirements.txt
streamlit
setuptools
wheel
pdfplumber
python-docx
python-pptx
pandas
pillow
sentence-transformers
scikit-learn
pytesseract
openai>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
tiktoken